fix flake error as suggested
