+++ minecraft/networking/packets/keep_alive_packet.py
@@ -1,0 +1,11 @@
+from .packet import Packet
+
+from minecraft.networking.types import (
+    VarInt
+)
+
+
+class KeepAlivePacket(Packet):
+    packet_name = "keep alive"
+    definition = [
+        {'keep_alive_id': VarInt}]
